--- RCAIDE/Library/Attributes/Gases/__init__.py
@@ -1,14 +1,0 @@
-# RCAIDE/Library/Attributes/Gases/__init__.py
-# (c) Copyright 2023 Aerospace Research Community LLC
-
-"""RCAIDE Package Setup
-"""
-
-# ----------------------------------------------------------------------------------------------------------------------
-#  IMPORT
-# ----------------------------------------------------------------------------------------------------------------------
-
-from .Gas    import Gas
-from .Air    import Air
-from .CO2    import CO2
-from .Steam  import Steam
