+++ RCAIDE/Library/Attributes/S_Propellants/Jet_A.py
@@ -13,6 +13,7 @@ from RCAIDE.Framework.Core import Data
 # ---------------------------------------------------------------------------------------------------------------------- 
 #  Jet_A1 Propellant Class
 # ----------------------------------------------------------------------------------------------------------------------  
+
 class Jet_A(Propellant):
     """Jet A class propellant
     """
