--- RCAIDE/Library/Components/Airfoils/__init__.py
@@ -1,12 +1,0 @@
-# RCAIDE/Library/Components/Airfoils/__init__.py
-# (c) Copyright 2023 Aerospace Research Community LLC
-
-"""RCAIDE Package Setup
-"""
-
-# ----------------------------------------------------------------------------------------------------------------------
-#  IMPORT
-# ----------------------------------------------------------------------------------------------------------------------
-
-from .Airfoil               import Airfoil
-from .NACA_4_Series_Airfoil import NACA_4_Series_Airfoil
