+++ RCAIDE/Library/Components/Energy/EnergyNetwork.py
@@ -1,0 +1,29 @@
+# RCAIDE/Library/Components/Energy/Network.py
+# (c) Copyright 2025 Aerospace Research Community LLC
+#
+# Created: Apr 2025, RCAIDE Team
+
+# ----------------------------------------------------------------------------------------------------------------------
+#  IMPORT
+# ----------------------------------------------------------------------------------------------------------------------
+
+from dataclasses import dataclass, field
+
+# package imports
+import numpy as np
+
+# RCAIDE imports
+import RCAIDE.Framework as rcf
+
+# ----------------------------------------------------------------------------------------------------------------------
+#  Network
+# ----------------------------------------------------------------------------------------------------------------------
+
+
+def _Network(State: rcf.State,
+                   Settings: rcf.Settings,
+                   System: rcf.System):
+
+
+                   
+    return State, Settings, System
