--- RCAIDE/Library/Components/Fuselages/__init__.py
@@ -1,13 +1,0 @@
-# RCAIDE/Library/Components/Fuselages/__init__.py
-# (c) Copyright 2023 Aerospace Research Community LLC
-
-"""RCAIDE Package Setup
-"""
-
-# ----------------------------------------------------------------------------------------------------------------------
-#  IMPORT
-# ----------------------------------------------------------------------------------------------------------------------
-
-from .Segment                    import Segment
-from .Fuselage                   import Fuselage
-from .Blended_Wing_Body_Fuselage import Blended_Wing_Body_Fuselage
