+++ RCAIDE/Library/Gases.py
@@ -1,0 +1,29 @@
+# RCAIDE/Library/Gases.py
+# (c) Copyright 2025 Aerospace Research Community LLC
+#
+# Created: Apr 2025, RCAIDE Team
+
+# ----------------------------------------------------------------------------------------------------------------------
+#  IMPORT
+# ----------------------------------------------------------------------------------------------------------------------
+
+from dataclasses import dataclass, field
+
+# package imports
+import numpy as np
+
+# RCAIDE imports
+import RCAIDE.Framework as rcf
+
+# ----------------------------------------------------------------------------------------------------------------------
+#  Gases
+# ----------------------------------------------------------------------------------------------------------------------
+
+
+def Gases(state: rcf.State,
+            settings: rcf.Settings,
+            system: rcf.System):
+
+
+                   
+    return state, settings, system
